backend/requirements.txt
fastapi[all]
uvicorn
motor
pymongo
scikit-learn
pandas
numpy
python-dotenv
